+++ src/main/java/com/link/plushies/mod/plushies/BeePlushie.java
@@ -1,4 +1,4 @@
-package com.link.plushies.mod.Plushies;
+package com.link.plushies.mod.plushies;
 
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
 import net.minecraft.block.*;
@@ -12,9 +12,9 @@ import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.util.shape.VoxelShapes;
 import net.minecraft.world.BlockView;
 
-public class TurtlePlushie extends HorizontalFacingBlock {
-    public TurtlePlushie(Settings settings) {
-        super(FabricBlockSettings.of(Material.WOOL).nonOpaque().sounds(BlockSoundGroup.WOOL).strength(0.6f));
+public class BeePlushie extends HorizontalFacingBlock {
+    public BeePlushie() {
+        super(FabricBlockSettings.of(Material.WOOL).nonOpaque().sounds(BlockSoundGroup.WOOL).strength(0.7f));
         setDefaultState(this.stateManager.getDefaultState().with(Properties.HORIZONTAL_FACING, Direction.NORTH));
     }
 
@@ -28,13 +28,13 @@ public class TurtlePlushie extends HorizontalFacingBlock {
         Direction dir = state.get(FACING);
         switch(dir) {
             case NORTH:
-                return VoxelShapes.cuboid(0f, 0f, 0f, 1f, 0.41f, 1f);
+                return VoxelShapes.cuboid(0f, 0f, 0f, 1f, 0.55f, 1f);
             case SOUTH:
-                return VoxelShapes.cuboid(0f, 0f, 0f, 1f, 0.411f, 1f);
+                return VoxelShapes.cuboid(0f, 0f, 0f, 1f, 0.55111f, 1f);
             case EAST:
-                return VoxelShapes.cuboid(0f, 0f, 0f, 1f, 0.4111f, 1f);
+                return VoxelShapes.cuboid(0f, 0f, 0f, 1f, 0.5511f, 1f);
             case WEST:
-                return VoxelShapes.cuboid(0f, 0f, 0f, 1f, 0.41111f, 1f);
+                return VoxelShapes.cuboid(0f, 0f, 0f, 1f, 0.551f, 1f);
             default:
                 return VoxelShapes.fullCube();
         }
