+++ src/main/java/com/link/plushies/mod/plushies/CatPlushie.java
@@ -1,4 +1,4 @@
-package com.link.plushies.mod.Plushies;
+package com.link.plushies.mod.plushies;
 
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
 import net.minecraft.block.*;
@@ -12,8 +12,8 @@ import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.util.shape.VoxelShapes;
 import net.minecraft.world.BlockView;
 
-public class PandaPlushie extends HorizontalFacingBlock {
-    public PandaPlushie(Settings settings) {
+public class CatPlushie extends HorizontalFacingBlock {
+    public CatPlushie() {
         super(FabricBlockSettings.of(Material.WOOL).nonOpaque().sounds(BlockSoundGroup.WOOL).strength(0.7f));
         setDefaultState(this.stateManager.getDefaultState().with(Properties.HORIZONTAL_FACING, Direction.NORTH));
     }
@@ -28,13 +28,13 @@ public class PandaPlushie extends HorizontalFacingBlock {
         Direction dir = state.get(FACING);
         switch(dir) {
             case NORTH:
-                return VoxelShapes.cuboid(0.2f, 0f, 0.325f, 0.8f, 0.6f, 0.7f);
+                return VoxelShapes.cuboid(0.35f, 0f, 0f, 0.7f, 0.55f, 1f);
             case SOUTH:
-                return VoxelShapes.cuboid(0.23f, 0f, 0.2f, 0.8f, 0.6f, 0.7f);
+                return VoxelShapes.cuboid(0.35f, 0f, 0f, 0.7f, 0.551f, 1f);
             case EAST:
-                return VoxelShapes.cuboid(0.22f, 0f, 0.3f, 0.7f, 0.6f, 0.8f);
+                return VoxelShapes.cuboid(0f, 0f, 0.35f, 1f, 0.55f, 0.7f);
             case WEST:
-                return VoxelShapes.cuboid(0.3f, 0f, 0.25f, 0.7f, 0.6f, 0.75f);
+                return VoxelShapes.cuboid(0f, 0f, 0.35f, 1f, 0.551f, 0.7f);
             default:
                 return VoxelShapes.fullCube();
         }
