+++ src/main/java/com/link/plushies/mod/plushies/DragonPlushie.java
@@ -1,4 +1,4 @@
-package com.link.plushies.mod.Plushies;
+package com.link.plushies.mod.plushies;
 
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
 import net.minecraft.block.*;
@@ -12,9 +12,9 @@ import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.util.shape.VoxelShapes;
 import net.minecraft.world.BlockView;
 
-public class SlimePlushie extends HorizontalFacingBlock {
-    public SlimePlushie(Settings settings) {
-        super(FabricBlockSettings.of(Material.WOOL).nonOpaque().sounds(BlockSoundGroup.WOOL).strength(0.6f));
+public class DragonPlushie extends HorizontalFacingBlock {
+    public DragonPlushie() {
+        super(FabricBlockSettings.of(Material.WOOL).nonOpaque().sounds(BlockSoundGroup.WOOL).strength(0.7f));
         setDefaultState(this.stateManager.getDefaultState().with(Properties.HORIZONTAL_FACING, Direction.NORTH));
     }
 
@@ -28,13 +28,13 @@ public class SlimePlushie extends HorizontalFacingBlock {
         Direction dir = state.get(FACING);
         switch(dir) {
             case NORTH:
-                return VoxelShapes.cuboid(0.25f, 0f, 0.25f, 0.75f, 0.91f, 0.75f);
+                return VoxelShapes.cuboid(0f, 0f, 0f, 1f, 0.85f, 1f);
             case SOUTH:
-                return VoxelShapes.cuboid(0.25f, 0f, 0.25f, 0.75f, 0.92f, 0.75f);
+                return VoxelShapes.cuboid(0f, 0f, 0f, 1f, 0.85111f, 1f);
             case EAST:
-                return VoxelShapes.cuboid(0.25f, 0f, 0.25f, 0.75f, 0.9f, 0.75f);
+                return VoxelShapes.cuboid(0f, 0f, 0f, 1f, 0.8511f, 1f);
             case WEST:
-                return VoxelShapes.cuboid(0.25f, 0f, 0.25f, 0.75f, 0.93f, 0.75f);
+                return VoxelShapes.cuboid(0f, 0f, 0f, 1f, 0.851f, 1f);
             default:
                 return VoxelShapes.fullCube();
         }
